refactor(audit): resolve structural findings (#5547)
